Optimize hash anti/semi join
When the optimizer chooses hash join for anti/semi join, the inner table
to build the hash table is always the table/relation that doesn't
produce output, even the table/relation is much larger that the outer
table. This commit adds an option for hash join to allow the inner table
to produce output. The idea is so-called hash right anti/semi join, like
right out join to left out join.